update: privateAssetItem
