+++ src/styles/components.scss
@@ -122,6 +122,8 @@
 }
 
 
+// PrivateAssetItem
+
 .zkAddressTooltipWrapper{
     max-width: 150px;
     .el-tooltip__popper{
